Added login

+++ backend_api/src/app.py
@@ -4,6 +4,7 @@ from db import db
 from flask import Flask
 from db import Show
 from db import Genre
+from db import User
 from flask import request
 
 app = Flask(__name__)
@@ -26,23 +27,47 @@ def failure_response(message, code=404):
     return json.dumps({"error": message}), code
 
 
-@app.route("/")
-def bruh():
-    print("Hello World!")
+@app.route("/login/", methods=["POST"])
+def login():
+    body = json.loads(request.data)
+    lst = ["username"]
+    for field in lst:
+        if field not in body:
+            return failure_response("Necessary information not provided!", 400)
+    find_user = User.query.filter_by(username=body.get('username')).first()
+    if find_user is None:
+        find_user = User(username=body.get('username'))
+        db.session.add(find_user)
+        db.session.commit()
+    return success_response(
+        find_user.serialize()
+    )
 
 
 @app.route("/api/watchlist/")
 def get_shows():
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
     return success_response(
-        {"shows": sorted([s.serialize_watch() for s in Show.query.filter_by(is_plan_to_watch=False)],
+        {"shows": sorted([s.serialize_watch() for s in Show.query.filter_by(is_plan_to_watch=False, user_id=body.get("user_id"))],
                          key=lambda show: show['finished'])}
     )
 
 
 @app.route("/api/watchlist/releasesort/")
 def get_shows_by_release_date():
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
     return success_response(
-        {"shows": sorted([s.serialize_watch() for s in Show.query.filter_by(is_plan_to_watch=False)],
+        {"shows": sorted([s.serialize_watch() for s in Show.query.filter_by(is_plan_to_watch=False, user_id=body.get("user_id"))],
                          key=lambda show: show['year_released'])}
     )
 
@@ -58,17 +83,21 @@ def get_genres():
 @app.route("/api/watchlist/", methods=["POST"])
 def create_watchlist_show():
     body = json.loads(request.data)
-    lst = ["name", "year_released", "start_date", "finished", "genre"]
+    lst = ["name", "year_released", "start_date",
+           "finished", "genre", "user_id"]
     for field in lst:
         if field not in body:
             return failure_response("Necessary information not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
     find_genre = Genre.query.filter_by(name=body.get('genre')).first()
     if find_genre is None:
         find_genre = Genre(name=body.get('genre'))
         db.session.add(find_genre)
         db.session.commit()
     new_show = Show(name=body.get("name"), year_released=body.get("year_released"), start_date=body.get(
-        "start_date"), finished=body.get("finished"), genre_id=find_genre.id, is_plan_to_watch=False)
+        "start_date"), finished=body.get("finished"), genre_id=find_genre.id, is_plan_to_watch=False, user_id=body.get("user_id"))
     db.session.add(new_show)
     db.session.commit()
 
@@ -77,7 +106,14 @@ def create_watchlist_show():
 
 @app.route("/api/watchlist/<int:show_id>/")
 def get_show_by_id(show_id):
-    show = Show.query.filter_by(id=show_id).first()
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
+    show = Show.query.filter_by(
+        id=show_id, user_id=body.get("user_id")).first()
     if show is None or show.is_plan_to_watch:
         return failure_response("Show not found!")
     return success_response(show.serialize_watch())
@@ -85,7 +121,14 @@ def get_show_by_id(show_id):
 
 @app.route("/api/watchlist/<int:show_id>/", methods=["DELETE"])
 def delete_watchlist_show(show_id):
-    show = Show.query.filter_by(id=show_id).first()
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
+    show = Show.query.filter_by(
+        id=show_id, user_id=body.get("user_id")).first()
     if show is None or show.is_plan_to_watch:
         return failure_response("Show not found!")
     db.session.delete(show)
@@ -93,27 +136,37 @@ def delete_watchlist_show(show_id):
     return success_response(show.serialize_watch())
 
 
-
 @app.route("/api/planlist/")
 def get_shows_by_release_date_plan():
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
     return success_response(
-        {"shows": sorted([s.serialize_plan() for s in Show.query.filter_by(is_plan_to_watch=True)],
+        {"shows": sorted([s.serialize_plan() for s in Show.query.filter_by(is_plan_to_watch=True, user_id=body.get("user_id"))],
                          key=lambda show: show['year_released'])}
     )
+
 
 @app.route("/api/planlist/", methods=["POST"])
 def create_planlist_show():
     body = json.loads(request.data)
-    lst = ["name", "year_released", "genre"]
+    lst = ["name", "year_released", "genre", "user_id"]
     for field in lst:
         if field not in body:
             return failure_response("Necessary information not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
     find_genre = Genre.query.filter_by(name=body.get('genre')).first()
     if find_genre is None:
         find_genre = Genre(name=body.get('genre'))
         db.session.add(find_genre)
         db.session.commit()
-    new_show = Show(name=body.get("name"), year_released=body.get("year_released"), genre_id=find_genre.id, is_plan_to_watch=True)
+    new_show = Show(name=body.get("name"), year_released=body.get(
+        "year_released"), genre_id=find_genre.id, is_plan_to_watch=True, user_id=body.get("user_id"))
     db.session.add(new_show)
     db.session.commit()
 
@@ -122,7 +175,14 @@ def create_planlist_show():
 
 @app.route("/api/planlist/<int:show_id>/")
 def get_show_by_id_plan(show_id):
-    show = Show.query.filter_by(id=show_id).first()
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
+    show = Show.query.filter_by(
+        id=show_id, user_id=body.get("user_id")).first()
     if show is None or not show.is_plan_to_watch:
         return failure_response("Show not found!")
     return success_response(show.serialize_plan())
@@ -130,14 +190,34 @@ def get_show_by_id_plan(show_id):
 
 @app.route("/api/planlist/<int:show_id>/", methods=["DELETE"])
 def delete_planlist_show(show_id):
-    show = Show.query.filter_by(id=show_id).first()
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
+    show = Show.query.filter_by(
+        id=show_id, user_id=body.get("user_id")).first()
     if show is None or not show.is_plan_to_watch:
         return failure_response("Show not found!")
     db.session.delete(show)
     db.session.commit()
     return success_response(show.serialize_plan())
 
+
+@app.route("/deleteuser/", methods=["DELETE"])
+def delete_user():
+    body = json.loads(request.data)
+    if "user_id" not in body:
+        return failure_response("user_id not provided!", 400)
+    user = User.query.filter_by(id=body.get("user_id")).first()
+    if user is None:
+        return failure_response("User not found!")
+    db.session.delete(user)
+    db.session.commit()
+    return success_response(user.serialize())
+
+
 if __name__ == "__main__":
     port = int(os.environ.get("PORT", 5000))
     app.run(host="0.0.0.0", port=port)
-
